pyproject.toml
[project]
name = "ai-english-teacher"
version = "0.1.0"
description = ""
authors = [
    { name = "Adam Zhou", email = "summychou@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.21.1",
    "numpy>=2.2.6",
    "pyaudio>=0.2.14",
    "rich>=14.0.0",
    "taskgroup>=0.2.2",
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "ruff>=0.11.12",
]

[[tool.uv.index]]
name = "tencentcloud"
url = "https://mirrors.cloud.tencent.com/pypi/simple/"
#name = "huaweicloud"
#url = "https://repo.huaweicloud.com/repository/pypi/simple/"
default = true

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    "E501",
    "T201",
]
line-length = 256
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"
